+++ hosts/chi/home.nix
@@ -1,1 +1,7 @@
-{ ... }: {}
+{ config, ... }:
+
+{
+  programs.zsh.envExtra = ''
+    export TERMINFO_DIRS=${config.home.profileDirectory}/share/terminfo:/usr/share/terminfo
+  '';
+}